Move create_annotated_heatmap from obvspython/vis.py -> obvs/vis.py; remove obvspyton directory
